feat: vinculo das paginas de gurpo com a pagina de cadastro de imóvel

+++ src/main/java/br/com/invistaix/InvistaIX/repository/ImovelRepository.java
@@ -15,5 +15,5 @@ public interface ImovelRepository extends JpaRepository<ImovelModel, Long> {
 	@Query("SELECT i FROM ImovelModel i WHERE i.numeroMatricula = :matricula")
 	Optional<ImovelModel> findByMatricula(@Param("matricula") String matricula);
 	@Query("SELECT i FROM ImovelModel i where i.idGrupo = :idGrupo")
-	List<ImovelModel> findAllInGrupo(@Param("idGrupo") Integer idGrupo);
+	List<ImovelModel> findAllInGrupo(@Param("idGrupo") Long idGrupo);
 }
